+++ src/main/java/com/nemonotfound/nemosambience/NemosAmbience.java
@@ -1,6 +1,7 @@
 package com.nemonotfound.nemosambience;
 
 import com.nemonotfound.nemosambience.client.particle.ModParticles;
+import com.nemonotfound.nemosambience.tag.ModBlockTags;
 import net.fabricmc.api.ModInitializer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -14,5 +15,6 @@ public class NemosAmbience implements ModInitializer {
     public void onInitialize() {
         log.info("Thanks for using Nemo's Ambience!");
         ModParticles.registerParticles();
+        ModBlockTags.register();
     }
 }