+++ src/main/java/com/nemonotfound/nemosambience/client/particle/FallingLeavesParticle.java
@@ -12,21 +12,21 @@ import net.minecraft.client.world.ClientWorld;
 import net.minecraft.registry.Registries;
 import net.minecraft.util.math.BlockPos;
 
-@Environment(value=EnvType.CLIENT)
-public class LeavesParticle
-extends SpriteBillboardParticle {
+@Environment(value = EnvType.CLIENT)
+public class FallingLeavesParticle
+        extends SpriteBillboardParticle {
 
     private float rotSpeed;
     private final float particleRandom;
     private final float spinAcceleration;
 
-    public LeavesParticle(ClientWorld world, double x, double y, double z, SpriteProvider spriteProvider) {
+    public FallingLeavesParticle(ClientWorld world, double x, double y, double z, SpriteProvider spriteProvider) {
         super(world, x, y, z);
         float f;
         this.setSprite(spriteProvider.getSprite(this.random.nextInt(12), 12));
-        this.rotSpeed = (float)Math.toRadians(this.random.nextBoolean() ? -30.0 : 30.0);
+        this.rotSpeed = (float) Math.toRadians(this.random.nextBoolean() ? -30.0 : 30.0);
         this.particleRandom = this.random.nextFloat();
-        this.spinAcceleration = (float)Math.toRadians(this.random.nextBoolean() ? -5.0 : 5.0);
+        this.spinAcceleration = (float) Math.toRadians(this.random.nextBoolean() ? -5.0 : 5.0);
         this.maxAge = 300;
         this.gravityStrength = 7.5E-4f;
         this.scale = f = this.random.nextBoolean() ? 0.05f : 0.075f;
@@ -37,8 +37,11 @@ extends SpriteBillboardParticle {
     }
 
     private void setParticleColor(ClientWorld world) {
-        int color = getFoliageColor(world);
+        BlockPos pos = BlockPos.ofFloored(x, y, z);
+        Block leavesBlock = world.getBlockState(pos.up()).getBlock();
+        String leavesName = Registries.BLOCK.getId(leavesBlock).getPath();
 
+        int color = getFoliageColor(world, pos, leavesName);
         float red = ((color >> 16) & 0xFF) / 255.0f;
         float green = ((color >> 8) & 0xFF) / 255.0f;
         float blue = (color & 0xFF) / 255.0f;
@@ -46,22 +49,17 @@ extends SpriteBillboardParticle {
         this.setColor(red, green, blue);
     }
 
-    private int getFoliageColor(ClientWorld world) {
+    private int getFoliageColor(ClientWorld world, BlockPos pos, String leavesName) {
         if (world == null) {
             return FoliageColors.getDefaultColor();
         }
 
-        BlockPos pos = BlockPos.ofFloored(x, y, z);
-        Block leavesBlock = world.getBlockState(pos.up()).getBlock();
-        String leavesName = Registries.BLOCK.getId(leavesBlock).getPath();
+        return switch (leavesName) {
+            case "birch_leaves" -> FoliageColors.getBirchColor();
+            case "spruce_leaves" -> FoliageColors.getSpruceColor();
+            default -> BiomeColors.getFoliageColor(world, pos);
+        };
 
-        if (leavesName.equals("birch_leaves")) {
-            return FoliageColors.getBirchColor();
-        } else if (leavesName.equals("spruce_leaves")) {
-            return FoliageColors.getSpruceColor();
-        }
-
-        return BiomeColors.getFoliageColor(world, pos);
     }
 
     @Override
@@ -84,14 +82,14 @@ extends SpriteBillboardParticle {
         float g = Math.min(f / 300.0f, 1.0f);
         double d = Math.cos(Math.toRadians(this.particleRandom * 60.0f)) * 2.0 * Math.pow(g, 1.25);
         double e = Math.sin(Math.toRadians(this.particleRandom * 60.0f)) * 2.0 * Math.pow(g, 1.25);
-        this.velocityX += d * (double)0.0025f;
-        this.velocityZ += e * (double)0.0025f;
+        this.velocityX += d * (double) 0.0025f;
+        this.velocityZ += e * (double) 0.0025f;
         this.velocityY -= this.gravityStrength;
         this.rotSpeed += this.spinAcceleration / 20.0f;
         this.prevAngle = this.angle;
         this.angle += this.rotSpeed / 20.0f;
         this.move(this.velocityX, this.velocityY, this.velocityZ);
-        if (this.onGround || this.maxAge < 299 && (this.velocityX == 0.0 || this.velocityZ == 0.0)) {
+        if (this.maxAge < 299 && (this.velocityX == 0.0 || this.velocityZ == 0.0)) {
             this.markDead();
         }
         if (this.dead) {
@@ -100,6 +98,13 @@ extends SpriteBillboardParticle {
         this.velocityX *= this.velocityMultiplier;
         this.velocityY *= this.velocityMultiplier;
         this.velocityZ *= this.velocityMultiplier;
+
+        if (this.onGround) {
+            this.velocityX = 0;
+            this.velocityY = 0;
+            this.velocityZ = 0;
+            this.rotSpeed = 0;
+        }
     }
 }