+++ src/main/java/com/nemonotfound/nemosambience/mixin/LeavesBlockMixin.java
@@ -1,11 +1,9 @@
 package com.nemonotfound.nemosambience.mixin;
 
+import com.nemonotfound.nemosambience.tag.ModBlockTags;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.LeavesBlock;
 import net.minecraft.client.util.ParticleUtil;
-import net.minecraft.particle.DefaultParticleType;
-import net.minecraft.registry.Registries;
-import net.minecraft.registry.tag.BlockTags;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Direction;
 import net.minecraft.util.math.random.Random;
@@ -16,7 +14,7 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-import static com.nemonotfound.nemosambience.client.particle.ModParticles.*;
+import static com.nemonotfound.nemosambience.client.particle.ModParticles.FALLING_LEAVES_PARTICLE;
 
 
 @Mixin(LeavesBlock.class)
@@ -24,38 +22,24 @@ public class LeavesBlockMixin {
 
 	@Inject(at = @At("HEAD"), method = "randomDisplayTick")
 	private void init(BlockState state, World world, BlockPos pos, Random random, CallbackInfo ci) {
-		String leavesName = Registries.BLOCK.getId(state.getBlock()).getPath();
-		boolean isLeavesBlock = state.getRegistryEntry().isIn(BlockTags.LEAVES);
+		boolean shouldBlockDropLeaves = state.getRegistryEntry().isIn(ModBlockTags.DROPS_FALLING_LEAVES);
 
-		if (isLeavesBlock) {
-			generateLeafParticles(world, pos, random, leavesName);
-		}
-
-
-	}
-
-	@Unique
-	private void generateLeafParticles(World world, BlockPos pos, Random random, String leavesName) {
-		switch (leavesName) {
-			case "oak_leaves" -> generateParticles(world, pos, random, OAK_LEAVES_Particle);
-			case "dark_oak_leaves" -> generateParticles(world, pos, random, DARK_OAK_LEAVES_Particle);
-			case "birch_leaves" -> generateParticles(world, pos, random, BIRCH_LEAVES_Particle);
-			case "spruce_leaves" -> generateParticles(world, pos, random, SPRUCE_LEAVES_Particle);
-			case "jungle_leaves" -> generateParticles(world, pos, random, JUNGLE_LEAVES_Particle);
-			case "mangrove_leaves" -> generateParticles(world, pos, random, MANGROVE_LEAVES_Particle);
-			case "acacia_leaves" -> generateParticles(world, pos, random, ACACIA_LEAVES_Particle);
+		if (shouldBlockDropLeaves) {
+			generateParticles(world, pos, random);
 		}
 	}
 
 	@Unique
-	private void generateParticles(World world, BlockPos pos, Random random, DefaultParticleType particleType) {
+	private void generateParticles(World world, BlockPos pos, Random random) {
 		if (random.nextInt(20) == 0) {
 			BlockPos blockPos = pos.down();
 			BlockState blockState = world.getBlockState(blockPos);
+
 			if (LeavesBlock.isFaceFullSquare(blockState.getCollisionShape(world, blockPos), Direction.UP)) {
 				return;
 			}
-			ParticleUtil.spawnParticle(world, pos, random, particleType);
+
+			ParticleUtil.spawnParticle(world, pos, random, FALLING_LEAVES_PARTICLE);
 		}
 	}
 }