+++ Changelog.md
@@ -1,4 +1,9 @@
-# Changelog v1.2
+# Changelog v1.4
+
+## Additions
+- Added tag "drops_falling_leaves"
+  - If you want to add leaf particles to your leaf blocks, add your block to the tag
+  - Works only for grey texture, since the color will be set (will be adapted in the future)
 
 ## Changes
-- Particles now change color by biome
+- Falling leaves particles now stay on ground for a bit